Fix test ProjectNotificationTests.test_project_membership_created
When more than one test is run, the primary key is no longer one

This test was failing due to not finding the membership with pk=1. Setting
first() rather than get(pk=1) seems to cause the test to pass.

Searching by email could also be a good way to do this.